pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vln-python-event-superschema"
version = "0.0.1"
authors = [
    { name="Ronald Landheer-Cieslak" , email="rlc@vlinder.ca" },
]
description = "Event super-schema and supporting functions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "jsonschema"
]

[projects.urls]
Homepage = "https://github.com/VlinderSoftware/python-event-superschema"
Issues = "https://github.com/VlinderSoftware/python-event-superschema/issues"

[project.optional-dependencies]
dev = [
    "pre-commit",
    "pylint",
    "pytest",
    "pytest-cov",
    "python-jose[cryptography]",
]
ci = [
    "sphinx",
    "sphinx-markdown-builder",
    "sphinx-autodoc-typehints",
]
paper = []

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
pythonpath = [
    ".",
    "event_superschema",
]

[tool.hatch.build.targets.wheel]
packages = ["event_superschema"]
